+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased] (beta)
 
+### Fixed
+- Fix repo connect not working without a config file
+
 ## [v0.5.0] (beta) - 2023-03-13
 
 ### Added